+++ packages/connect-wallet-modal/src/helpers/userAgents.ts
@@ -1,7 +1,0 @@
-import { UAParser } from 'ua-parser-js';
-
-const parser = new UAParser();
-
-export const os = parser.getOS();
-export const isIOS = os.name === 'iOS';
-export const isAndroid = os.name === 'Android';