+++ packages/core-react/src/helpers/index.ts
@@ -1,4 +1,3 @@
 export * from './getUnsupportedChainError';
 export * from './checkTermsAccepted';
-export * from './userAgents';
 export * from './useLocalStorage';